feat(large): Refactor HR Zone Logic and Visualization (#6858)
