+++ hooks/useWorkoutSessionManager.ts
@@ -6,10 +6,12 @@ import {
   WorkoutSessionData,
   HrDataPoint,
 } from '../lib/workout-session-storage'
-import { HrZoneName } from '../lib/shared/hr-zones'
+import {
+  HrZoneName,
+  calculateHrZone,
+  calculateMaxHr,
+} from '@/lib/shared/hr-zones'
 import { v4 as uuidv4 } from 'uuid'
-import { calculateHrZone } from '../lib/hrm/zones'
-import { calculateMaxHr } from '@/lib/shared/hr-zones'
 import { useAppSnackbar } from './useAppSnackbar'
 import { isSessionStale } from '../lib/workout-session'
