+++ utils/visualization.ts
@@ -8,8 +8,7 @@ import { WorkoutData } from '../types'
 import { WorkoutItem } from '../types/workout'
 import { WorkoutColumnsProps } from '@/components/WorkoutColumns'
 import theme from '../lib/theme'
-import { calculateHrZone } from '../lib/hrm/zones'
-import { HrZoneName } from '../lib/shared/hr-zones'
+import { calculateHrZone, HrZoneName } from '../lib/shared/hr-zones'
 
 // Define types for MUI color props
 type MuiColor =
@@ -26,6 +25,7 @@ type HrZoneUi = {
   color: string
   progressColor: string
   bgColor: string
+  textColor: string
 }
 
 export const HR_ZONE_UI_PROPS_MAP: Record<HrZoneName, HrZoneUi> = {
@@ -33,37 +33,44 @@ export const HR_ZONE_UI_PROPS_MAP: Record<HrZoneName, HrZoneUi> = {
     color: 'text-blue-400',
     progressColor: theme.palette.secondary.main,
     bgColor: theme.palette.secondary.main,
+    textColor: theme.palette.getContrastText(theme.palette.secondary.main),
   },
   [HrZoneName.FatBurn]: {
     color: 'text-green-500',
     progressColor: theme.palette.success.main,
     bgColor: theme.palette.success.main,
+    textColor: '#FFFFFF',
   },
   [HrZoneName.Cardio]: {
     color: 'text-yellow-500',
     progressColor: theme.palette.warning.dark,
     bgColor: theme.palette.warning.dark,
+    textColor: '#FFFFFF',
   },
   [HrZoneName.Peak]: {
     color: 'text-red-500',
     progressColor: theme.palette.primary.main,
     bgColor: theme.palette.primary.main,
+    textColor: theme.palette.getContrastText(theme.palette.primary.main),
   },
   [HrZoneName.Max]: {
     color: 'text-purple-600',
     progressColor: '#9333ea',
     bgColor: '#9C27B0',
+    textColor: theme.palette.getContrastText('#9C27B0'),
   },
-  // Add placeholder properties for non-displayable zones
+  // Fallback UI properties for non-calculable zones (e.g., missing data)
   [HrZoneName.NoData]: {
     color: 'text-gray-400',
     progressColor: '#9ca3af',
     bgColor: '#B0BEC5', // Lighter grey for better visibility
+    textColor: '#FFFFFF',
   },
   [HrZoneName.Unknown]: {
     color: 'text-gray-400',
     progressColor: '#9ca3af',
     bgColor: '#9ca3af',
+    textColor: '#FFFFFF',
   },
 }
 
@@ -102,25 +109,14 @@ export const getHrZoneProps = (
   // 2. Look up the UI properties from the map
   const zoneUiProps = HR_ZONE_UI_PROPS_MAP[zoneName]
 
-  // 3. Determine text color - force white for specific zones for better contrast
-  let textColor = theme.palette.getContrastText(zoneUiProps.bgColor)
-  if (
-    zoneName === HrZoneName.NoData ||
-    zoneName === HrZoneName.Unknown ||
-    zoneName === HrZoneName.FatBurn ||
-    zoneName === HrZoneName.Cardio
-  ) {
-    textColor = '#FFFFFF' // Force white text for grey, green, and yellow zones
-  }
-
-  // 4. Combine domain data with UI properties
+  // 3. Combine domain data with UI properties
   return {
     zone: zoneName, // The enum member is a string at runtime
     percentage: percentage,
     color: zoneUiProps.color,
     progressColor: zoneUiProps.progressColor,
     backgroundColor: zoneUiProps.bgColor,
-    textColor: textColor,
+    textColor: zoneUiProps.textColor,
     bpm: bpm,
   }
 }
